requirements.txt
alembic==1.0.0
amqp==2.3.2
asn1crypto==0.24.0
beautifulsoup4==4.6.0
billiard==3.5.0.4
blinker==1.4
celery==4.2.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.1.4
distro==1.3.0
dominate==2.3.1
feedparser==5.2.1
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-Script-Extras==0.1.4
Flask-SQLAlchemy==2.3.2
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
itsdangerous==0.24
jeepney==0.3.1
jikan-sqlalchemy-utils==0.0.7
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
kombu==4.2.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
migrate==0.3.8
oauth==1.0.1
oauthlib==2.0.6
olefile==0.45.1
pbr==4.2.0
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
py-postgresql==1.2.1
pycairo==1.16.2
pycparser==2.18
pycrypto==2.6.1
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-dateutil==2.7.3
python-debian==0.1.32
python-editor==1.0.3
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
SQLAlchemy==1.2.10
SQLAlchemy-Utils==0.33.3
testresources==2.0.1
urllib3==1.22
vine==1.1.4
visitor==0.1
wadllib==1.3.2
webencodings==0.5
Werkzeug==0.14.1
zope.interface==4.3.2
